pysrc/requirements.txt
dune-client==0.1.1
psycopg2-binary==2.9.3
python-dotenv>=0.20.0
requests>=2.28.1
pandas==1.5.0
pandas-stubs==1.5.1.221024
ndjson>=0.3.1


black==22.6.0
mypy==0.982
pylint==2.14.4
pytest==7.1.2